+++ java-exam-evaluator-annotations/src/main/java/hu/sed/evaluator/annotation/syntax/ConstructorCheck.java
@@ -30,5 +30,5 @@ public @interface ConstructorCheck {
      *
      * @return maximum score
      */
-    int score() default 1;
+    double score() default 1;
 }
